requirements.txt
aiofiles==23.2.1
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2024.2.0
ftfy==6.2.0
h11==0.14.0
idna==3.7
intel-openmp==2021.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
tinycss2==1.3.0
torch>=1.8.0
torchaudio>=1.8.0
torchvision>=0.9.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
